refactor: replace typer with cyclopts
