+++ pyproject.toml
@@ -38,17 +38,13 @@ version = "2.0.3"
 
 [project.optional-dependencies]
 cli = [
-    "async-mega-py[default]",
-    "typer-slim>=0.21.1"
-]
-default = [
-    "python-dotenv>=1.2.1",
-    "rich>=14.0.0"
+    "cyclopts>=4.10.1",
+    "python-dotenv>=1.2.1"
 ]
 
 [project.scripts]
-async-mega-py = "mega.cli:main"
-mega-py = "mega.cli:main"
+async-mega-py = "mega.__main__:app"
+mega-py = "mega.__main__:app"
 
 [project.urls]
 Homepage = "https://github.com/NTFSvolume/mega.py"
@@ -131,6 +127,5 @@ dev = [
     "pytest-asyncio>=0.25.0",
     "pytest>=8.3.5",
     "rich>=14.0.0",
-    "ruff>=0.15.0",
-    "typer-slim>=0.21.1"
+    "ruff>=0.15.0"
 ]
